authml/backend/requirements.txt
Flask
flask-cors
pymongo
redis
python-dotenv
PyJWT
werkzeug
rdkit
joblib
scikit-learn
numpy
seaborn 
matplotlib